+++ traits_futures/tests/common_future_tests.py
@@ -48,8 +48,10 @@ COMPLETE_VALID_SEQUENCES = {
     "SX",
     "CSR",
     "CSX",
+    "CSA",
     "SCR",
     "SCX",
+    "SCA",
     "CA",
 }